requirements.txt
bcrypt==4.3.0
dlib==12.22.99
flask==3.0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
pillow==10.4.0
pillow-heif==0.18.0
pyotp==2.9.0
sqlalchemy==2.0.43